+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 All notable changes to this "vscode-msr" extension will be documented in this file.
 
+## [2.1.88]
+
+- Supported copying function/script body to `aliasBody` using `to-alias-body`; Auto convert `Tab`/`spaces` + `return`/`exit`.
+
 ## [2.1.87]
 
-- Improved multi-line `to-alias-body` + supported `to-alias-body` on Linux/MacOS.
+- Improved multi-lines `to-alias-body` + supported `to-alias-body` on Linux/MacOS.
 
 ## [2.1.86]